+++ sdk/computelimit/azure-resourcemanager-computelimit/src/main/java/com/azure/resourcemanager/computelimit/ComputeLimitManager.java
@@ -30,10 +30,12 @@ import com.azure.resourcemanager.computelimit.implementation.FeaturesImpl;
 import com.azure.resourcemanager.computelimit.implementation.GuestSubscriptionsImpl;
 import com.azure.resourcemanager.computelimit.implementation.OperationsImpl;
 import com.azure.resourcemanager.computelimit.implementation.SharedLimitsImpl;
+import com.azure.resourcemanager.computelimit.implementation.VmFamiliesImpl;
 import com.azure.resourcemanager.computelimit.models.Features;
 import com.azure.resourcemanager.computelimit.models.GuestSubscriptions;
 import com.azure.resourcemanager.computelimit.models.Operations;
 import com.azure.resourcemanager.computelimit.models.SharedLimits;
+import com.azure.resourcemanager.computelimit.models.VmFamilies;
 import java.time.Duration;
 import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
@@ -54,6 +56,8 @@ public final class ComputeLimitManager {
     private SharedLimits sharedLimits;
 
     private Features features;
+
+    private VmFamilies vmFamilies;
 
     private final ComputeLimitClient clientObject;
 
@@ -316,6 +320,18 @@ public final class ComputeLimitManager {
             this.features = new FeaturesImpl(clientObject.getFeatures(), this);
         }
         return features;
+    }
+
+    /**
+     * Gets the resource collection API of VmFamilies.
+     * 
+     * @return Resource collection API of VmFamilies.
+     */
+    public VmFamilies vmFamilies() {
+        if (this.vmFamilies == null) {
+            this.vmFamilies = new VmFamiliesImpl(clientObject.getVmFamilies(), this);
+        }
+        return vmFamilies;
     }
 
     /**