+++ sdk/computelimit/azure-resourcemanager-computelimit/src/main/java/com/azure/resourcemanager/computelimit/fluent/ComputeLimitClient.java
@@ -73,4 +73,11 @@ public interface ComputeLimitClient {
      * @return the FeaturesClient object.
      */
     FeaturesClient getFeatures();
+
+    /**
+     * Gets the VmFamiliesClient object to access its operations.
+     * 
+     * @return the VmFamiliesClient object.
+     */
+    VmFamiliesClient getVmFamilies();
 }