+++ sdk/computelimit/azure-resourcemanager-computelimit/src/main/java/com/azure/resourcemanager/computelimit/fluent/FeaturesClient.java
@@ -125,4 +125,60 @@ public interface FeaturesClient {
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
     OperationStatusResultInner enable(String location, String featureName, Context context);
+
+    /**
+     * Disables a compute limit feature for the subscription at the specified location.
+     * 
+     * @param location The name of the Azure region.
+     * @param featureName The name of the Feature.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return the {@link SyncPoller} for polling of long-running operation.
+     */
+    @ServiceMethod(returns = ReturnType.LONG_RUNNING_OPERATION)
+    SyncPoller<PollResult<OperationStatusResultInner>, OperationStatusResultInner> beginDisable(String location,
+        String featureName);
+
+    /**
+     * Disables a compute limit feature for the subscription at the specified location.
+     * 
+     * @param location The name of the Azure region.
+     * @param featureName The name of the Feature.
+     * @param context The context to associate with this operation.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return the {@link SyncPoller} for polling of long-running operation.
+     */
+    @ServiceMethod(returns = ReturnType.LONG_RUNNING_OPERATION)
+    SyncPoller<PollResult<OperationStatusResultInner>, OperationStatusResultInner> beginDisable(String location,
+        String featureName, Context context);
+
+    /**
+     * Disables a compute limit feature for the subscription at the specified location.
+     * 
+     * @param location The name of the Azure region.
+     * @param featureName The name of the Feature.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return the response.
+     */
+    @ServiceMethod(returns = ReturnType.SINGLE)
+    OperationStatusResultInner disable(String location, String featureName);
+
+    /**
+     * Disables a compute limit feature for the subscription at the specified location.
+     * 
+     * @param location The name of the Azure region.
+     * @param featureName The name of the Feature.
+     * @param context The context to associate with this operation.
+     * @throws IllegalArgumentException thrown if parameters fail the validation.
+     * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
+     * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
+     * @return the response.
+     */
+    @ServiceMethod(returns = ReturnType.SINGLE)
+    OperationStatusResultInner disable(String location, String featureName, Context context);
 }