+++ sdk/computelimit/azure-resourcemanager-computelimit/src/main/java/com/azure/resourcemanager/computelimit/implementation/ComputeLimitClientImpl.java
@@ -31,6 +31,7 @@ import com.azure.resourcemanager.computelimit.fluent.FeaturesClient;
 import com.azure.resourcemanager.computelimit.fluent.GuestSubscriptionsClient;
 import com.azure.resourcemanager.computelimit.fluent.OperationsClient;
 import com.azure.resourcemanager.computelimit.fluent.SharedLimitsClient;
+import com.azure.resourcemanager.computelimit.fluent.VmFamiliesClient;
 import java.io.IOException;
 import java.lang.reflect.Type;
 import java.nio.ByteBuffer;
@@ -186,6 +187,20 @@ public final class ComputeLimitClientImpl implements ComputeLimitClient {
     }
 
     /**
+     * The VmFamiliesClient object to access its operations.
+     */
+    private final VmFamiliesClient vmFamilies;
+
+    /**
+     * Gets the VmFamiliesClient object to access its operations.
+     * 
+     * @return the VmFamiliesClient object.
+     */
+    public VmFamiliesClient getVmFamilies() {
+        return this.vmFamilies;
+    }
+
+    /**
      * Initializes an instance of ComputeLimitClient client.
      * 
      * @param httpPipeline The HTTP pipeline to send requests through.
@@ -202,11 +217,12 @@ public final class ComputeLimitClientImpl implements ComputeLimitClient {
         this.defaultPollInterval = defaultPollInterval;
         this.endpoint = endpoint;
         this.subscriptionId = subscriptionId;
-        this.apiVersion = "2026-03-20";
+        this.apiVersion = "2026-04-30";
         this.operations = new OperationsClientImpl(this);
         this.guestSubscriptions = new GuestSubscriptionsClientImpl(this);
         this.sharedLimits = new SharedLimitsClientImpl(this);
         this.features = new FeaturesClientImpl(this);
+        this.vmFamilies = new VmFamiliesClientImpl(this);
     }
 
     /**