+++ sdk/computelimit/azure-resourcemanager-computelimit/CHANGELOG.md
@@ -2,11 +2,18 @@
 
 ## 1.0.0 (2026-04-21)
 
+- Azure Resource Manager ComputeLimit client library for Java. This package contains Microsoft Azure SDK for ComputeLimit Management SDK. Microsoft Azure Compute Limit Resource Provider. Package api-version 2026-04-30. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
 - Azure Resource Manager ComputeLimit client library for Java. This package contains Microsoft Azure SDK for ComputeLimit Management SDK. Microsoft Azure Compute Limit Resource Provider. Package api-version 2026-03-20. For documentation on how to use this package, please see [Azure Management Libraries for Java](https://aka.ms/azsdk/java/mgmt).
 
 ### Features Added
 
 * `models.OperationStatusResult` was added
+
+* `models.VmFamilyProperties` was added
+
+* `models.VmFamilies` was added
+
+* `models.VmFamily` was added
 
 * `models.FeatureProperties` was added
 
@@ -19,6 +26,7 @@
 #### `ComputeLimitManager` was modified
 
 * `features()` was added
+* `vmFamilies()` was added
 
 ## 1.0.0-beta.1 (2025-11-12)